Move tasklist-specific questions to tasklistHelper
