+++ server/utils/taskListUtils.ts
@@ -1,10 +1,40 @@
+import type { FormSections, FormSection } from '@approved-premises/ui'
 import { ApprovedPremisesApplication } from '../@types/shared'
 import { JourneyType, Task } from '../@types/ui'
 import applyPaths from '../paths/apply'
 import assessPaths from '../paths/assess'
-import { previousTaskIsComplete } from './applicationUtils'
+import Apply from '../form-pages/apply'
 
-const taskLinkHelper = (task: Task, application: ApprovedPremisesApplication, journeyType?: JourneyType): string => {
+const taskIds = Object.keys(Apply.pages)
+
+const taskIsComplete = (task: Task, application: ApprovedPremisesApplication): boolean => {
+  return application.data?.[task.id]
+}
+
+const previousTaskIsComplete = (task: Task, application: ApprovedPremisesApplication): boolean => {
+  const previousTaskId = taskIds[taskIds.indexOf(task.id) - 1]
+  return previousTaskId ? application.data?.[previousTaskId] : true
+}
+
+const getTaskStatus = (task: Task, application: ApprovedPremisesApplication): string => {
+  if (!taskIsComplete(task, application) && !previousTaskIsComplete(task, application)) {
+    return `<strong class="govuk-tag govuk-tag--grey app-task-list__tag" id="${task.id}-status">Cannot start yet</strong>`
+  }
+
+  if (!taskIsComplete(task, application) && previousTaskIsComplete(task, application)) {
+    return `<strong class="govuk-tag govuk-tag--grey app-task-list__tag" id="${task.id}-status">Not started</strong>`
+  }
+
+  return `<strong class="govuk-tag app-task-list__tag" id="${task.id}-status">Completed</strong>`
+}
+
+const getCompleteSectionCount = (sections: FormSections, application: ApprovedPremisesApplication): number => {
+  return sections.filter((section: FormSection) => {
+    return section.tasks.filter((task: Task) => taskIsComplete(task, application)).length === section.tasks.length
+  }).length
+}
+
+const taskLink = (task: Task, application: ApprovedPremisesApplication, journeyType?: JourneyType): string => {
   if (previousTaskIsComplete(task, application)) {
     const firstPage = Object.keys(task.pages)[0]
 
@@ -26,4 +56,4 @@ const taskLinkHelper = (task: Task, application: ApprovedPremisesApplication, jo
   return task.title
 }
 
-export default taskLinkHelper
+export { taskLink, getCompleteSectionCount, getTaskStatus, previousTaskIsComplete, taskIsComplete }
