+++ server/views/assessments/show.njk
@@ -27,9 +27,9 @@
               {% for task in section.tasks %}
                 <li class="app-task-list__item">
                   <span class="app-task-list__task-name">
-                    {{ taskLink(task, assessment, 'assessments')  }}
+                    {{ TasklistUtils.taskLink(task, assessment, 'assessments') | safe   }}
                   </span>
-                  {{ getTaskStatus(task, assessment) }}
+                  {{ TasklistUtils.getTaskStatus(task, assessment) | safe  }}
                 </li>
               {% endfor %}
             </ul>
